# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Profo 房产数据中心",
    "description": "轻量级、本地化、高性能的房产数据仓库系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/cashflow": {
      "post": {
        "tags": [
          "cashflow",
          "cashflow"
        ],
        "summary": "Create Cashflow Record",
        "description": "创建现金流记录",
        "operationId": "create_cashflow_record_api_v1_projects__project_id__cashflow_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashFlowRecordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cashflow",
          "cashflow"
        ],
        "summary": "Get Project Cashflow",
        "description": "获取项目现金流明细和汇总",
        "operationId": "get_project_cashflow_api_v1_projects__project_id__cashflow_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/cashflow/{record_id}": {
      "delete": {
        "tags": [
          "cashflow",
          "cashflow"
        ],
        "summary": "Delete Cashflow Record",
        "description": "删除现金流记录",
        "operationId": "delete_cashflow_record_api_v1_projects__project_id__cashflow__record_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "项目ID",
              "title": "Project Id"
            },
            "description": "项目ID"
          },
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "记录ID",
              "title": "Record Id"
            },
            "description": "记录ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/monitor/communities/{community_id}/sentiment": {
      "get": {
        "tags": [
          "monitor"
        ],
        "summary": "Get Sentiment",
        "operationId": "get_sentiment_api_v1_monitor_communities__community_id__sentiment_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Community Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketSentimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitor/communities/{community_id}/trends": {
      "get": {
        "tags": [
          "monitor"
        ],
        "summary": "Get Trends",
        "operationId": "get_trends_api_v1_monitor_communities__community_id__trends_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Community Id"
            }
          },
          {
            "name": "months",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 24,
              "minimum": 1,
              "default": 6,
              "title": "Months"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrendData"
                  },
                  "title": "Response Get Trends Api V1 Monitor Communities  Community Id  Trends Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitor/ai-strategy": {
      "post": {
        "tags": [
          "monitor"
        ],
        "summary": "Generate Strategy",
        "operationId": "generate_strategy_api_v1_monitor_ai_strategy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIStrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIStrategyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/monitor/communities/{community_id}/radar": {
      "get": {
        "tags": [
          "monitor"
        ],
        "summary": "Get Neighborhood Radar",
        "description": "获取周边竞品雷达数据，包含分渠道统计",
        "operationId": "get_neighborhood_radar_api_v1_monitor_communities__community_id__radar_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Community Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborhoodRadarResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{community_id}/competitors": {
      "get": {
        "tags": [
          "communities"
        ],
        "summary": "Get Competitors",
        "operationId": "get_competitors_api_v1_communities__community_id__competitors_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Community Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompetitorResponse"
                  },
                  "title": "Response Get Competitors Api V1 Communities  Community Id  Competitors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "communities"
        ],
        "summary": "Add Competitor",
        "operationId": "add_competitor_api_v1_communities__community_id__competitors_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Community Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCompetitorRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communities/{community_id}/competitors/{competitor_id}": {
      "delete": {
        "tags": [
          "communities"
        ],
        "summary": "Remove Competitor",
        "operationId": "remove_competitor_api_v1_communities__community_id__competitors__competitor_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "community_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Community Id"
            }
          },
          {
            "name": "competitor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Competitor Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIStrategyRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "user_context": {
            "type": "string",
            "title": "User Context"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "user_context"
        ],
        "title": "AIStrategyRequest"
      },
      "AIStrategyResponse": {
        "properties": {
          "report_markdown": {
            "type": "string",
            "title": "Report Markdown"
          },
          "risk_points": {
            "$ref": "#/components/schemas/RiskPoints"
          },
          "action_plan": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Action Plan"
          }
        },
        "type": "object",
        "required": [
          "report_markdown",
          "risk_points",
          "action_plan"
        ],
        "title": "AIStrategyResponse"
      },
      "AddCompetitorRequest": {
        "properties": {
          "competitor_community_id": {
            "type": "integer",
            "title": "Competitor Community Id"
          }
        },
        "type": "object",
        "required": [
          "competitor_community_id"
        ],
        "title": "AddCompetitorRequest"
      },
...
      "CashFlowCategory": {
        "type": "string",
        "enum": [
          "履约保证金",
          "中介佣金",
          "装修费",
          "营销费",
          "其他支出",
          "税费",
          "运营费",
          "回收保证金",
          "溢价款",
          "服务费",
          "其他收入",
          "售房款"
        ],
        "title": "CashFlowCategory",
        "description": "现金流分类枚举"
      },
      "CashFlowRecordCreate": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/CashFlowType"
          },
          "category": {
            "$ref": "#/components/schemas/CashFlowCategory"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "related_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Stage"
          }
        },
        "type": "object",
        "required": [
          "type",
          "category",
          "amount",
          "date"
        ],
        "title": "CashFlowRecordCreate",
        "description": "创建现金流"
      },
      "CashFlowRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "record_date": {
            "type": "string",
            "format": "date-time",
            "title": "Record Date",
            "description": "发生日期"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "备注"
          },
          "operator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Id",
            "description": "经办人ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date",
            "readOnly": true
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "readOnly": true
          },
          "related_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Stage",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "type",
          "category",
          "amount",
          "record_date",
          "created_at",
          "updated_at",
          "date",
          "description",
          "related_stage"
        ],
        "title": "CashFlowRecordResponse",
        "description": "现金流记录响应 - 适配新的FinanceRecord表"
      },
      "CashFlowResponse": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/CashFlowRecordResponse"
            },
            "type": "array",
            "title": "Records"
          },
          "summary": {
            "$ref": "#/components/schemas/CashFlowSummary"
          }
        },
        "type": "object",
        "required": [
          "records",
          "summary"
        ],
        "title": "CashFlowResponse"
      },
      "CashFlowSummary": {
        "properties": {
          "total_income": {
            "type": "string",
            "title": "Total Income"
          },
          "total_expense": {
            "type": "string",
            "title": "Total Expense"
          },
          "net_cash_flow": {
            "type": "string",
            "title": "Net Cash Flow"
          },
          "roi": {
            "type": "number",
            "title": "Roi"
          },
          "annualized_return": {
            "type": "number",
            "title": "Annualized Return",
            "default": 0.0
          },
          "holding_days": {
            "type": "integer",
            "title": "Holding Days",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "total_income",
          "total_expense",
          "net_cash_flow",
          "roi"
        ],
        "title": "CashFlowSummary"
      },
      "CashFlowType": {
        "type": "string",
        "enum": [
          "income",
          "expense"
        ],
        "title": "CashFlowType",
        "description": "现金流类型枚举"
      },
...
      "CompetitorResponse": {
        "properties": {
          "community_id": {
            "type": "integer",
            "title": "Community Id"
          },
          "community_name": {
            "type": "string",
            "title": "Community Name"
          },
          "avg_price": {
            "type": "number",
            "title": "Avg Price"
          },
          "on_sale_count": {
            "type": "integer",
            "title": "On Sale Count"
          }
        },
        "type": "object",
        "required": [
          "community_id",
          "community_name",
          "avg_price",
          "on_sale_count"
        ],
        "title": "CompetitorResponse"
      },
...
      "FloorStats": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "deals_count": {
            "type": "integer",
            "title": "Deals Count"
          },
          "deal_avg_price": {
            "type": "number",
            "title": "Deal Avg Price"
          },
          "current_count": {
            "type": "integer",
            "title": "Current Count"
          },
          "current_avg_price": {
            "type": "number",
            "title": "Current Avg Price"
          }
        },
        "type": "object",
        "required": [
          "type",
          "deals_count",
          "deal_avg_price",
          "current_count",
          "current_avg_price"
        ],
        "title": "FloorStats"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketSentimentResponse": {
        "properties": {
          "floor_stats": {
            "items": {
              "$ref": "#/components/schemas/FloorStats"
            },
            "type": "array",
            "title": "Floor Stats"
          },
          "inventory_months": {
            "type": "number",
            "title": "Inventory Months"
          }
        },
        "type": "object",
        "required": [
          "floor_stats",
          "inventory_months"
        ],
        "title": "MarketSentimentResponse"
      },
...
      "NeighborhoodRadarItem": {
        "properties": {
          "community_id": {
            "type": "integer",
            "title": "Community Id"
          },
          "community_name": {
            "type": "string",
            "title": "Community Name"
          },
          "is_subject": {
            "type": "boolean",
            "title": "Is Subject"
          },
          "listing_count": {
            "type": "integer",
            "title": "Listing Count"
          },
          "listing_beike": {
            "type": "integer",
            "title": "Listing Beike"
          },
          "listing_iaij": {
            "type": "integer",
            "title": "Listing Iaij"
          },
          "listing_avg_price": {
            "type": "number",
            "title": "Listing Avg Price"
          },
          "deal_count": {
            "type": "integer",
            "title": "Deal Count"
          },
          "deal_beike": {
            "type": "integer",
            "title": "Deal Beike"
          },
          "deal_iaij": {
            "type": "integer",
            "title": "Deal Iaij"
          },
          "deal_avg_price": {
            "type": "number",
            "title": "Deal Avg Price"
          },
          "spread_percent": {
            "type": "number",
            "title": "Spread Percent"
          },
          "spread_label": {
            "type": "string",
            "title": "Spread Label"
          }
        },
        "type": "object",
        "required": [
          "community_id",
          "community_name",
          "is_subject",
          "listing_count",
          "listing_beike",
          "listing_iaij",
          "listing_avg_price",
          "deal_count",
          "deal_beike",
          "deal_iaij",
          "deal_avg_price",
          "spread_percent",
          "spread_label"
        ],
        "title": "NeighborhoodRadarItem",
        "description": "周边竞品雷达单项数据"
      },
      "NeighborhoodRadarResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NeighborhoodRadarItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "NeighborhoodRadarResponse",
        "description": "周边竞品雷达响应"
      },
...
      "RiskPoints": {
        "properties": {
          "profit_critical_price": {
            "type": "number",
            "title": "Profit Critical Price"
          },
          "daily_cost": {
            "type": "number",
            "title": "Daily Cost"
          }
        },
        "type": "object",
        "required": [
          "profit_critical_price",
          "daily_cost"
        ],
        "title": "RiskPoints"
      },
...
      "TrendData": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "listing_price": {
            "type": "number",
            "title": "Listing Price"
          },
          "deal_price": {
            "type": "number",
            "title": "Deal Price"
          },
          "volume": {
            "type": "integer",
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "month",
          "listing_price",
          "deal_price",
          "volume"
        ],
        "title": "TrendData"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}